pyproject.toml
[project]
name = "ModularBot"
version = "0.1.0"
description = "This is an modular bot discord"
readme = "README.md"
requires-python = ">=3.12"
author = "lleans"
classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "discord-py>=2.4.0",
    "easy-pil>=0.4.0",
    "iso639-lang>=2.5.1",
    "pytz>=2024.2",
    "wavelink>=3.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ModularBot"]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "ruff>=0.11.0",
]

[tool.ruff]
include = ["ModularBot/*", "bot.py", "config.py"]
preview = true
target-version = "py312"

[tool.ruff.format]
indent-style = "tab"
docstring-code-format = true